attempts to get a udp tunnel going, but no answer
It seems we don't get the ServerSync before there has been an actual UDP connection.
Tried to emulate this by sending an UDP ping via UdpTunnel, or a dummy voice packet via the same mechanism.
ping seems to be ignored in UdpTunnel in murmur's code, and our voice message is probably malformed.

Easiest way would probably be to send the ping via UDP, but need to figure out that encryption first.